requirements.txt
langchain
langgraph
openai
tiktoken
pinecone-client
python-dotenv
arxiv
google-search-results
pydantic
rich
langchain-openai
wikipedia
reportlab
streamlit